cv-inference/requirements.txt
ultralytics>=8.2.0
onnxruntime-gpu>=1.18.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
